Revise thread-safety logic
